+++ CHANGELOG.md
@@ -4,3 +4,7 @@
 - Draft FastApi code and sqlalchemy code
 - Currently mock data is pulled through a dummy open source API (to be changed after FastAPI is implemented)
 
+
+22 Sep 2024
+- FastAPI works with sqlite and no response models
+- routing, crud, db initialization and models setup